Fix problem with an incorrect implicit cast to float

+++ ete2/coretype/tree.py
@@ -1440,7 +1440,7 @@ class TreeNode(object):
                 result.pop()
             (lo, hi, end) = (mids[0], mids[-1], len(result))
             prefixes = [PAD] * (lo+1) + [PA+'|'] * (hi-lo-1) + [PAD] * (end-hi)
-            mid = (lo + hi) / 2
+            mid = int((lo + hi) / 2)
             prefixes[mid] = char1 + '-'*(LEN-2) + prefixes[mid][-1]
             result = [p+l for (p,l) in zip(prefixes, result)]
             if show_internal:
